Some changes(yeah I know, very informative... But I'm too tired rn so whatever...
