+++ assets/properties/commands.properties
@@ -1,3 +1,6 @@
+botUnknownCmdError="\
+Неизвестная команда. Игнорирую.
+"
 botGreetingCmd="\
 Здравствуйте, я - EBot.\
 Для получения информации о функционале введите команду `/help`\
@@ -9,11 +12,11 @@ botHelpCmd="\
     /help - показать доступный функционал\
     /email - посмотреть свою электронную почту\
     /setemail - установить электронную почту\
-    /gdfolderid - посмотреть ID папки бота на GD\
-    /setgdfolderid - установить ID папки бота на GD\
-    /credentials - посмотреть свои GD Credentials\
-    /setcredentials - установить GD Credentials\
-    /gdguide - посмотреть инструкцию для подключения GD
+    /gdfolderid - посмотреть ID папки бота на GoogleDrive\
+    /setgdfolderid - установить ID папки бота на Google Drive\
+    /credentials - посмотреть свои Google Drive Credentials\
+    /setcredentials - установить Google Drive Credentials\
+    /gdguide - посмотреть инструкцию по подключению Google Drive\
 "
 
 botHelpForwardCmd="\
@@ -24,12 +27,14 @@ botHelpForwardCmd="\
 botHelpWhatIfErrorCmd="\
 Если при выполнении какой-либо команды произойдёт ошибка\
 (о чём вам напишет бот), то вы вернётесь в исходное состояние\
-до запуска каких-либо команд и вам нужно будет вызвать эту команду заново\
+до запуска каких-либо команд и вам нужно будет вызвать эту команду заново.\
 "
-
+botHelpWhatIfCmdInsideCmd="\
+Аналогично запуск команд внутри других приведёт к сбросу состояния.\
+"
 botHelpGoogleDiskCmd="\
 Если вы хотите привязать Гугл-Диск для отправки вложений(картинок, документов и т.д.),\
-то вам потребуется привязать GD Credentials с помощью \
+то вам потребуется привязать Google Drive Credentials с помощью \
 команды `/setcredentials`\
 "
 
@@ -56,13 +61,13 @@ botSetGDFolderIDExtraInfoCmd="\
 Заметка: Вы можете отослать либо ID папки на гугл-диске, либо просто URL этой папки.\
 "
 botUserInvalidCredentialsError="\
-Ошибка: Ваши GD Credentials невалидны(%s).\
+Ошибка: Ваши Google Drive Credentials невалидны(%s).\
 Вы можете изменить их при помощи команды /setcredentials.\
 "
 botUserSetterSuccessCmd="\
 Ваше новое значение %s:\
 "
-botSetGDCredentialsExtraInfo0="Отправьте GD Client Secrets"
+botSetGDCredentialsExtraInfo0="Отправьте Google Drive Client Secrets"
 botSetGDCredentialsExtraInfo1="\
 Заметка: что такое "Google Disk Client Secrets" и как их получить вы можете узнать\
 по команде /gdguide.\
@@ -74,6 +79,6 @@ botSetGDCredentialsExtraInfo2="\
 botSetGDCredentialsExtraInfo3="Пожалуйста, перейдите по этой ссылке и подтвердите допуск:"
 botSetGDCredentialsExtraInfo4="Отправьте код подтверждения отдельным сообщением"
 botHumanEmail="e-mail"
-botHumanGDClientSecrets="GD Client Secrets"
-botHumanGDCredentials="GD Credentials"
-botHumanGDFolderID="ID папки на GD"
+botHumanGDClientSecrets="Google Drive Client Secrets"
+botHumanGDCredentials="Google Drive Credentials"
+botHumanGDFolderID="ID папки на Google Drive"
