+++ assets/properties/general.properties
@@ -1,5 +1,5 @@
 botEmail="emailbottelegram@yandex.ru"
-botLetterSendPreamble="Ошибка при отправке письма:"
+botLetterSendErrorPreamble="Ошибка при отправке письма:"
 botLetterSendInvalidEmailErrorMsg='\
 Ваш e-mail("%s") невалиден.\
 Вы можете изменить свою почту при помощи команды `/setemail`.\
@@ -8,5 +8,4 @@ botGDClientSecretsInvalidErrorMsg="\
 Ваши client secrets невалидны(%s).
 Вы можете изменить свои Client Secrets пр помощи команды `/setcredentials`
 "
-googleCloudAccountGuidePath="assets/GDConsoleGuide.md"
 tempFilesPath="assets/tmp"
