+++ assets/properties/google.properties
@@ -1,6 +1,9 @@
+googleCloudAccountGuidePath="assets/GDConsoleGuide.md"
 googleOauth2Scope="https://www.googleapis.com/auth/drive"
 googleOAuth2UnsupportedFlowErr="This OAuth 2.0 flow is unsupported:"
 googleDiskFilePrefix="https://drive.google.com/file/d/"
+googleService="drive"
+googleServiceVersion='v3'
 GDFileUploadSuccess="Файл успешно загружен!"
 GDFileUploadMaybeError="Загрузка может быть неуспешна:"
 GDFileUploadFolderNotExist="\
